test(FeedbackPopup): merge tests into one file
